requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-filter==2.0.0
django-model-utils==3.1.2
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
lxml==4.2.5
mysqlclient==1.3.13
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
whitenoise==4.1
